\name{Burn}
\docType{data}
\alias{Burn}
\title{Data from Table 3 in GallantBruckPrickett(2006)}
\description{
   This is an .rda file containing a mixture-process variable experiment with 3 mixture components and one process variable. The data
   has been modified from the original article so that the sum of the tree mixure components is 1.0.
}
\usage{data(SneeMq)}
\format{ An 15 x 11 data frame}
\source{source}
\references{ 
    Gallant, F.M., Bruck, H.A. and Prickett, S.E. (2006) Effects of Twin-Screw Extrusion Processing on the Burning Rate of Composite Propellants, \emph{Propellants, Explosives, Pyrotechnics}, Vol. 31, pp 456-465.
}
\keyword{datasets}