\name{conmx}
\docType{data}
\alias{conmx}
\title{Example constraint matrix from Piepel 1988}
\description{
   This is an .rda file containing the constraint matrix.
}
\usage{data(conmx)}
\format{ An 8 x 4 matrix}
\source{source}
\references{ 
    Piepel, G. F. (1988) Programs for Generating Extreme Vertices and Centroids of Linearly Constrained Experimental Regions, \emph{Journal of Quality Technology}, Vol. 20, No. 2.
}
\keyword{datasets}