\name{etch}
\docType{data}
\alias{etch}
\title{Data from Etch rate experiment in Table 12.4 of Myers and Montgomery(2002) }
\description{
   This is an .rda file containing a mixture experiment with 3 mixture components. The mixture components are x1, x2, and x3. The response is erate. 
}
\usage{data(etch)}
\format{ An 14 x 4 data frame}
\source{source}
\references{ 
    Myers, R. H. and Montgomery D. C. (2002) \emph{Response Surface Methodology - Product and Process Optimization Using Designed Experiments} John Wiley and Sons, New York.
}
\keyword{datasets}