\name{fishp}
\docType{data}
\alias{fishp}
\title{Data from Cornell's famous fish patty mixture process variable experiment}
\description{
   This is an .rda file design and response.
}
\usage{data(fishp)}
\format{ An 56 x 7 data frame}
\source{source}
\references{ 
  Cornell, J. A., \emph{Experiments with Mixtures, Third Edition},  John Wiley and Sons, 2002, pp 361-365.
}
\keyword{datasets}