\name{SneeMq}
\docType{data}
\alias{SneeMq}
\title{Data from Snee and Marquart's Screening Experiment with contreined mixture components}
\description{
   This is an .rda file design and response.
}
\usage{data(SneeMq)}
\format{ An 16 x 9 data frame}
\source{source}
\references{ 
    Snee, D. D. and Marquardt D. W. (1976) Screening Concepts and designs for experiments with mixtures, \emph{Technometrics}, Vol. 18, pp 19-29.
}
\keyword{datasets}