\name{Burn}
\docType{data}
\alias{Burn}
\title{Data from Table 4 in Gallant,Prickett, Cesarec, and Bruck(2008)}
\description{
   This is an .rda file containing a mixture-process variable experiment with 3 mixture components and one process variable. z is the coded value of RPM, and y is average Burning rate for test presure 500(psig) 
}
\usage{data(Burn)}
\format{ An 15 x 5 data frame}
\source{source}
\references{ 
    Gallant, F.M., Prickett, S.E. Cesarec, M. and Bruck, H.A.(2008) Ingredients and processing effects on burning rates of composite rocket propellants utilizing a reduced-run mixture-process experiment design, \emph{Chemometrics and intellegent laboratory systems}, Vol. 90, pp 49-63.
}
\keyword{datasets}